+++ server/src/modules/pages/db/entity/PageDetailsEntity.ts
