+++ server/src/modules/pages/db/entity/index.ts
@@ -1,0 +1,2 @@
+export type { PageEntity } from "./PageEntity";
+export type { PageDetailsEntity } from "./PageDetailsEntity";
